test: move provider mocks to core package
Providers are defined in the core package, so it's the appropriate place to store mocks,
making them accessible for other packages. Moves the `key-management` mock to the correct context.

also fix some mock data types
