+++ packages/core/package.json
@@ -53,6 +53,7 @@
   "devDependencies": {
     "@cardano-sdk/util-dev": "^0.9.0",
     "@types/lodash": "^4.14.182",
+    "delay": "^5.0.0",
     "eslint": "^7.32.0",
     "jest": "^28.1.3",
     "madge": "^5.0.1",
