+++ packages/key-management/test/mocks/index.ts
@@ -1,2 +1,3 @@
 export * from './mockKeyAgentDependencies';
+export * from './stakeKey';
 export * from './testKeyAgent';
