+++ packages/key-management/test/mocks/stakeKey.ts
@@ -1,0 +1,6 @@
+import { KeyRole } from '../../src';
+
+export const stakeKeyDerivationPath = {
+  index: 0,
+  role: KeyRole.Stake
+};
