+++ packages/wallet/test/hardware/trezor/TrezorKeyAgent.integration.test.ts
@@ -12,7 +12,7 @@ import {
   mockRewardsProvider,
   mockTxSubmitProvider,
   mockUtxoProvider
-} from '../../mocks';
+} from '../../../../core/test/mocks';
 
 const createWallet = async (keyAgent: KeyAgent) => {
   const txSubmitProvider = mockTxSubmitProvider();
