+++ packages/wallet/test/integration/SupplyDistributionTrackerTrigger.test.ts
@@ -2,7 +2,7 @@ import { createInMemorySupplyDistributionStores } from '../../src/persistence';
 import { createSupplyDistributionTracker } from '../../src';
 import { createWallet } from './util';
 import { dummyLogger } from 'ts-log';
-import { mockNetworkInfoProvider } from '../mocks';
+import { mockNetworkInfoProvider } from '../../../core/test/mocks';
 
 describe('SupplyDistributionTracker', () => {
   it('accepts an ObservableWallet as "trigger"', async () => {
